Arreglo bbdd

+++ backend/database/migrations/2025_06_08_220743_create_habitacionesreservadas_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->date('fecha_entrada');
             $table->date('fecha_salida');
             $table->string('comida');
-            $table->integer('pagado');
+            $table->decimal('pagado', 10, 2);
             $table->string('tipo_pago');
             $table->timestamps();
         });
